Cut out object initialize method from DataFormatter
